various improvements :
Ensured cardholderName contains at least 2 words and is limited to 30 characters.
Displayed cardholderName in uppercase on the card.
Validated month and year fields with a regex.
Refined error message logic for month and year (format, length, value).
Increased the minimum length of the card number from 12 to 16 digits.
Displayed the card number in the format XXXX XXXX XXXX XXXX